scan

+++ database/migrations/2026_09_01_092559_add_processed_to_scan_uids_table.php
@@ -9,16 +9,20 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('scan_uids', function (Blueprint $table) {
-            $table->boolean('processed')
-                ->default(false)
-                ->after('uid');
+            if (!Schema::hasColumn('scan_uids', 'processed')) {
+                $table->boolean('processed')
+                    ->default(false)
+                    ->after('uid');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('scan_uids', function (Blueprint $table) {
-            $table->dropColumn('processed');
+            if (Schema::hasColumn('scan_uids', 'processed')) {
+                $table->dropColumn('processed');
+            }
         });
     }
 };
